refactor(LotteryPurchase): 이름에 맞지 않는 패키지 제거 후 클래스 이동
