+++ src/test/java/lotterymachine/vo/LotteryPurchaseTest.java
@@ -1,10 +1,9 @@
 package lotterymachine.vo;
 
-import org.assertj.core.api.Assertions;
+import lotterymachine.LotteryPurchase;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class LotteryPurchaseTest {
